(TAB-26/ACPE-003) Prepare by shifting content validation to its own source file. No functional changes.
